requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.24
botocore==1.23.24
celery==5.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.5
django-eb-sqs-worker==0.5.0
django-static-fontawesome==5.14.0.0
django-storages==1.12.3
docutils==0.18.1
fontawesome-free==5.15.3
gevent==1.4.0
greenlet==0.4.15
idna==3.2
jmespath==0.10.0
kombu==5.2.2
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pycparser==2.19
pycurl==7.43.0.5
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
redis==4.0.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
